+++ Sources/MisticaCommon/Colors/O2ColorPalette.swift
@@ -92,7 +92,7 @@ struct O2Colors: MisticaColors {
     let textPrimary = palette.grey6 | palette.grey2
     let textPrimaryInverse = palette.white | palette.grey2
     let textSecondary = palette.grey5 | palette.grey4
-    let textSecondaryInverse = palette.white | palette.grey4
+    let textSecondaryInverse = palette.o2BluePrimary15 | palette.grey4
 
     // STATES
     let error = palette.pepper