+++ Sources/MisticaCommon/Fonts/FontStyle.swift
@@ -21,7 +21,6 @@ import UIKit
     case textPreset8
     case textPreset9
     case textPreset10
-    case cardTitlePreset
 
     func preferredFont(
         weight: Font.Weight,
@@ -76,8 +75,6 @@ import UIKit
             return "TextPreset9"
         case .textPreset10:
             return "TextPreset10"
-        case .cardTitlePreset:
-            return "CardTitlePreset"
         }
     }
 }
@@ -128,8 +125,6 @@ private extension FontStyle {
             return 32
         case .textPreset10:
             return 32
-        case .cardTitlePreset:
-            return 18
         }
     }
 
@@ -141,8 +136,7 @@ private extension FontStyle {
             return .subheadline
         case .textPreset3:
             return .body
-        case .textPreset4,
-             .cardTitlePreset:
+        case .textPreset4:
             return .headline
         case .textPreset5:
             return .title3
@@ -165,8 +159,7 @@ private extension FontStyle {
             return .subheadline
         case .textPreset3:
             return .body
-        case .textPreset4,
-             .cardTitlePreset:
+        case .textPreset4:
             return .headline
         case .textPreset5:
             if #available(iOS 14.0, *) {