+++ Sources/MisticaCommon/MisticaConfig.swift
@@ -12,12 +12,13 @@ public enum MisticaConfig {
     public private(set) static var currentColors: MisticaColors = MovistarColors()
     public private(set) static var currentBrandAssets: MisticaBrandAssets = DefaultMisticaBrandAssets()
     public private(set) static var currentStyledControls = [MisticaControlStyle]()
+    public private(set) static var currentFontWeights: MisticaFontWeights = MovistarFontWeights()
 
     // MARK: Public Setup
 
     public static var brandStyle: BrandStyle = .movistar {
         didSet {
-            configureColorPalette(for: brandStyle)
+            configure(for: brandStyle)
             MisticaAppearance.setUp(controls: currentStyledControls)
         }
     }
@@ -29,20 +30,24 @@ public enum MisticaConfig {
 }
 
 private extension MisticaConfig {
-    static func configureColorPalette(for _: BrandStyle) {
+    static func configure(for _: BrandStyle) {
         switch brandStyle {
         case .movistar:
             currentColors = MovistarColors()
             currentBrandAssets = DefaultMisticaBrandAssets()
+            currentFontWeights = MovistarFontWeights()
         case .vivo:
             currentColors = VivoColors()
             currentBrandAssets = DefaultMisticaBrandAssets()
+            currentFontWeights = VivoFontWeights()
         case .o2:
             currentColors = O2Colors()
             currentBrandAssets = DefaultMisticaBrandAssets()
+            currentFontWeights = O2FontWeights()
         case .blau:
             currentColors = BlauColors()
             currentBrandAssets = DefaultMisticaBrandAssets()
+            currentFontWeights = BlauFontWeights()
         case .custom(let colors, let assets):
             currentColors = colors
             currentBrandAssets = assets