+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [23.3.0](https://github.com/Telefonica/mistica-ios/compare/v23.2.0...v23.3.0) (2023-04-25)
+
+
+### Features
+
+* **Colors:** Update textSecondaryInverse for all brands ([ac37ba6](https://github.com/Telefonica/mistica-ios/commit/ac37ba619849226157f531eab9b90214043eb272))
+
+# [23.2.0](https://github.com/Telefonica/mistica-ios/compare/v23.1.0...v23.2.0) (2023-04-18)
+
+
+### Features
+
+* **FontWeight:** Added brandeable FontWeights ([d4280c2](https://github.com/Telefonica/mistica-ios/commit/d4280c24f4af605aa788c3a7f8390ea0a180f81d))
+
+# [23.1.0](https://github.com/Telefonica/mistica-ios/compare/v23.0.2...v23.1.0) (2023-04-13)
+
+
+### Features
+
+* **TitleView:** IOS-8290 default title style based on the brand ([#268](https://github.com/Telefonica/mistica-ios/issues/268)) ([7d3a347](https://github.com/Telefonica/mistica-ios/commit/7d3a347cf46bd57510f23172e8c40a38f72c801e))
+
 ## [23.0.2](https://github.com/Telefonica/mistica-ios/compare/v23.0.1...v23.0.2) (2023-03-21)